# backend/openapi.yaml
openapi: 3.0.3
info:
  title: DevDex API - Stack Overflow Style Q&A Platform
  description: API documentation for DevDex - A developer Q&A platform with authentication, questions, answers, voting, tags, and payment integration
  version: 2.0.0
  contact:
    name: DevDex Support
servers:
  - url: http://localhost:5000
    description: Local development server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in HTTP-only cookie
  schemas:
    User:
      type: object
      required:
        - username
        - email
        - password
      properties:
        _id:
          type: string
          description: MongoDB ObjectId
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          writeOnly: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    Question:
      type: object
      required:
        - title
        - desc
        - username
        - userId
      properties:
        _id:
          type: string
          readOnly: true
        title:
          type: string
          description: Question title
        desc:
          type: string
          description: Question body/description
        photo:
          type: string
          description: Optional image URL
        username:
          type: string
        userId:
          type: string
        categories:
          type: array
          items:
            type: string
          description: Tags associated with the question
        votes:
          type: integer
          default: 0
        upvoters:
          type: array
          items:
            type: string
        downvoters:
          type: array
          items:
            type: string
        acceptedAnswerId:
          type: string
          nullable: true
        views:
          type: integer
          default: 0
        answerCount:
          type: integer
          default: 0
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    Answer:
      type: object
      required:
        - body
        - questionId
        - userId
        - username
      properties:
        _id:
          type: string
          readOnly: true
        body:
          type: string
          description: Answer content
        questionId:
          type: string
        userId:
          type: string
        username:
          type: string
        votes:
          type: integer
          default: 0
        upvoters:
          type: array
          items:
            type: string
        downvoters:
          type: array
          items:
            type: string
        isAccepted:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    Tag:
      type: object
      properties:
        _id:
          type: string
          readOnly: true
        name:
          type: string
          description: Unique tag name (lowercase)
        description:
          type: string
        questionCount:
          type: integer
          default: 0
        createdAt:
          type: string
          format: date-time
          readOnly: true
...
    VoteRequest:
      type: object
      required:
        - voteType
        - userId
      properties:
        voteType:
          type: integer
          enum: [1, -1]
          description: 1 for upvote, -1 for downvote
        userId:
          type: string
...
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '500':
          description: Server error
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
        '400':
          description: User not found
        '401':
          description: Wrong password
...
  /api/auth/refetch:
    get:
      tags: [Authentication]
      summary: Refetch user data from token
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User data retrieved

  # User Endpoints
...
  /api/posts:
    get:
      tags: [Questions]
      summary: Get all questions with filtering and sorting
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search in title and description
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, active, unanswered, votes]
        - name: tag
          in: query
          schema:
            type: string
          description: Filter by tag name
      responses:
        '200':
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
...
  /api/answers/create:
    post:
      tags: [Answers]
      summary: Create a new answer
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body, questionId, userId, username]
              properties:
                body:
                  type: string
                questionId:
                  type: string
                userId:
                  type: string
                username:
                  type: string
      responses:
        '200':
          description: Answer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Answer'
  /api/answers/question/{questionId}:
    get:
      tags: [Answers]
      summary: Get answers for a question
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of answers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Answer'
  /api/answers/{id}:
    put:
      tags: [Answers]
      summary: Update answer
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: string
      responses:
        '200':
          description: Answer updated

    delete:
      tags: [Answers]
      summary: Delete answer
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Answer deleted
  /api/answers/{id}/vote:
    put:
      tags: [Answers]
      summary: Vote on an answer
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteRequest'
      responses:
        '200':
          description: Vote registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Answer'

  # Tag Endpoints
  /api/tags:
    get:
      tags: [Tags]
      summary: Get all tags
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'

    post:
      tags: [Tags]
      summary: Create or update tag
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Tag created/updated
...
  /api/upload:
    post:
      tags: [Upload]
      summary: Upload an image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Image uploaded

  # Payment Endpoint
...